- updated to the complete to have exception handling in the launched thread

+++ Installers/RockInstaller/RockInstaller/Complete.aspx
@@ -167,19 +167,20 @@
         DeleteDirectory( serverPath + @"\bin" );
 
         // move the rock application into place
-        DirectoryContentsMove( serverPath + @"\rock", serverPath );
-        
-        // delete this page
-        File.Delete( serverPath + @"Complete.aspx" );
+        if ( DirectoryContentsMove( serverPath + @"\rock", serverPath ) )
+        {
+            // delete this page
+            File.Delete( serverPath + @"Complete.aspx" );
 
-        // delete a web.config if it already exists in the root, this is not the rock one
-        File.Copy( serverPath + @"\webconfig.xml", serverPath + @"\web.config", true );
+            // delete a web.config if it already exists in the root, this is not the rock one
+            File.Copy( serverPath + @"\webconfig.xml", serverPath + @"\web.config", true );
 
-        // delete web config template
-        File.Delete( serverPath + @"\webconfig.xml" );
-        
-        // delete rock install directory
-        Directory.Delete( serverPath + @"\rock", true );
+            // delete web config template
+            File.Delete( serverPath + @"\webconfig.xml" );
+
+            // delete rock install directory
+            Directory.Delete( serverPath + @"\rock", true );
+        }
     }
 
     private void DeleteDirectory( string target_dir )
@@ -201,57 +202,79 @@
         Directory.Delete( target_dir, false );
     }
 
-    private void DirectoryContentsMove( string sourceDirName, string destDirName )
+    private bool DirectoryContentsMove( string sourceDirName, string destDirName )
     {
         // this will move the contents of a folder into the contents of another
         // if a folder already exists in the directory with the same name it
         // will be deleted. ONLY USE THIS IF THIS USE CASE WORKS FOR YOU!!!
-        
-        DirectoryInfo sourceDirectory = new DirectoryInfo( sourceDirName );
-        DirectoryInfo[] sourceChildDirectories = sourceDirectory.GetDirectories();
 
-        // If the source directory does not exist, throw an exception.
-        if ( !sourceDirectory.Exists )
+        try
         {
-            throw new DirectoryNotFoundException(
-                "Source directory does not exist or could not be found: "
-                + sourceDirName );
-        }
+            DirectoryInfo sourceDirectory = new DirectoryInfo( sourceDirName );
+            DirectoryInfo[] sourceChildDirectories = sourceDirectory.GetDirectories();
 
-        // If the destination directory does not exist, create it.
-        if ( !Directory.Exists( destDirName ) )
-        {
-            Directory.CreateDirectory( destDirName );
-        }
-
-        // move child directories
-        foreach (DirectoryInfo dir in sourceChildDirectories)
-        {
-
-            string destChildDirectory = Path.Combine( destDirName, dir.Name );
-            if ( Directory.Exists( destChildDirectory ) )
+            // If the source directory does not exist, throw an exception.
+            if ( !sourceDirectory.Exists )
             {
-                DeleteDirectory( destChildDirectory );
-            }
-            dir.MoveTo( destChildDirectory );
-        }
-        
-        // move child files
-        FileInfo[] files = sourceDirectory.GetFiles();
-
-        foreach(FileInfo file in files)
-        {
-            // Create the path to the new copy of the file.
-            string temppath = Path.Combine( destDirName, file.Name );
-
-            // check if file exists in dest if so delete
-            if ( File.Exists( temppath ) )
-            {
-                File.Delete( temppath );
+                throw new DirectoryNotFoundException(
+                    "Source directory does not exist or could not be found: "
+                    + sourceDirName );
             }
 
-            // move the file.
-            file.MoveTo( temppath );
+            // If the destination directory does not exist, create it.
+            if ( !Directory.Exists( destDirName ) )
+            {
+                Directory.CreateDirectory( destDirName );
+            }
+
+            // move child directories
+            foreach ( DirectoryInfo dir in sourceChildDirectories )
+            {
+                string destChildDirectory = Path.Combine( destDirName, dir.Name );
+                if ( Directory.Exists( destChildDirectory ) )
+                {
+                    DeleteDirectory( destChildDirectory );
+                }
+                dir.MoveTo( destChildDirectory );
+            }
+
+            // move child files
+            FileInfo[] files = sourceDirectory.GetFiles();
+
+            foreach ( FileInfo file in files )
+            {
+                // Create the path to the new copy of the file.
+                string temppath = Path.Combine( destDirName, file.Name );
+
+                // check if file exists in dest if so delete
+                if ( File.Exists( temppath ) )
+                {
+                    File.Delete( temppath );
+                }
+
+                // move the file.
+                file.MoveTo( temppath );
+            }
+        }
+        
+        catch (Exception ex)
+        {
+            LogException( ex );
+            return false;
+        }
+        
+        return true;
+    }
+    
+    private void LogException(Exception ex)
+    {
+        string directory = AppDomain.CurrentDomain.BaseDirectory;
+        string filePath = Path.Combine( directory, "InstallExceptions.csv" );
+        string when = DateTime.Now.ToString();
+        while ( ex != null )
+        {
+            File.AppendAllText( filePath, string.Format( "{0},{1},\"{2}\"\r\n", when, ex.GetType(), ex.Message ) );
+            ex = ex.InnerException;
         }
     }
